app/requirements.txt
Django==1.9.5
gunicorn==19.4.5
Pillow==3.2.0
django-taggit==0.18.1
